add task_6

+++ ru.appline/src/main/java/task_5/logic/Add.java
@@ -2,5 +2,5 @@ package task_5.logic;
 
 public interface Add {
     /** Returns a+b */
-    float add(float a, float b);
+    int add(int a, int b);
 }
