+++ ru.appline/src/main/java/task_5/logic/Divide.java
@@ -2,5 +2,5 @@ package task_5.logic;
 
 public interface Divide {
     /** Returns a/b */
-    float divide(float a, float b);
+    int divide(int a, int b);
 }
