+++ ru.appline/src/main/java/task_5/logic/Multiply.java
@@ -2,5 +2,5 @@ package task_5.logic;
 
 public interface Multiply {
     /** Returns a*b */
-    float multiply(float a, float b);
+    int multiply(int a, int b);
 }
