+++ ru.appline/src/main/java/task_5/logic/Subtract.java
@@ -2,5 +2,5 @@ package task_5.logic;
 
 public interface Subtract {
     /** Returns a-b */
-    float subtract(float a, float b);
+    int subtract(int a, int b);
 }
